requirements/requirements.txt
rasa==3.6.13
jieba==0.42.1
python-dotenv==1.0.0
transformers==4.29.1

langchain==0.0.333
typing-inspect==0.8.0
typing_extensions==4.5.0
tiktoken==0.4.0
unstructured==0.7.3
pdf2image==1.16.3
sentence_transformers==2.2.2

faiss-cpu==1.7.4
pypdfium2==4.18.0

whoosh==2.7.4
trafilatura==1.6.1
docx2txt==0.8

pyautogen==0.2.0b3
jenkinsapi==0.3.13